%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{aroma.affymetrix.tools-package}
\alias{aroma.affymetrix.tools-package}
\alias{aroma.affymetrix.tools}
\docType{package}

\title{Package aroma.affymetrix.tools}


 \description{
  Methods that simplify the setup of S3 generic functions
and S3 methods.  Major effort has been made in making
definition of methods as simple as possible with a minimum of
maintainance for package developers.  For example, generic
functions are created automatically, if missing, and name
conflict are automatically solved, if possible.  The method
setMethodS3() is a good start for those who in the future want
to migrate to S4.  This is a cross-platform package implemented
in pure R and is generating standard S3 methods.

   This package should be considered to be in an alpha or beta phase.
   You should expect the API to be changing over time.
 }

 \section{Requirements}{
   This package requires \pkg{aroma.affymetrix.tools}.
 }

 \section{Installation and updates}{

   To install this package, see instructions at
   \url{http://www.braju.com/R/}.
 }

 \author{
   Mark Robinson <mrobinson@wehi.edu.au>
 }

 \section{License}{
   The releases of this package is licensed under
   LGPL version 2.1 or newer.
 }

 \references{
 }
\keyword{package}
